feat: added reusable component InputSearch
